optimize cache performance
linked list have poor performance as they grow large (>100000).  this is
caused by a significant amount of cache misses.  to fix this we
allocated a large array for the data and had the linked list point to
the begining of the data structures.  we still have O(1) insertion and
deletion.